Crate matrix_sdk_test

Modules

test_json

Test data for the matrix-sdk crates.

Structs

EventBuilder

The EventBuilder struct can be used to easily generate valid sync responses for testing. These can be then fed into either Client or Room.

InvitedRoomBuilder

JoinedRoomBuilder

LeftRoomBuilder

Enums

EphemeralTestEvent

Test events that can be added to the ephemeral events.

GlobalAccountDataTestEvent

Test events that can be added to the global account data.

PresenceTestEvent

Test events that can be added to the presence events.

RoomAccountDataTestEvent

Test events that can be added to the room account data.

StateTestEvent

Test events that can be added to the state.

StrippedStateTestEvent

Test events that can be added to the stripped state.

SyncResponseFile

Embedded sync response files

TimelineTestEvent

Test events that can be added to the timeline.

Functions

bulk_room_members

Create m.room.member events in the given range.

response_from_file

sync_response

Get specific API responses for testing

Attribute Macros

async_test

Attribute to use wasm_bindgen_test for wasm32 targets and tokio::test for everything else with async-support and custom result-types