Module matrix_sdk_test::test_json::api_responses

Responses to client API calls.

Statics

DEVICES

GET /_matrix/client/v3/devices

GET_ALIAS

GET /_matrix/client/v3/directory/room/{roomAlias}

KEYS_QUERY

POST /_matrix/client/v3/keys/query

KEYS_QUERY_TWO_DEVICES_ONE_SIGNED

POST /_matrix/client/v3/keys/query For a set of 2 devices own by a user named web2. First device is unsigned, second one is signed

KEYS_UPLOAD

``

LOGIN

Successful call to POST /_matrix/client/v3/login without auto-discovery.

LOGIN_RESPONSE_ERR

Failed call to POST /_matrix/client/v3/login

LOGIN_TYPES

GET /_matrix/client/v3/login

LOGIN_WITH_DISCOVERY

Successful call to POST /_matrix/client/v3/login with auto-discovery.

LOGIN_WITH_REFRESH_TOKEN

Successful call to POST /_matrix/client/v3/login with a refresh token.

NOT_FOUND

Failed call to an endpoint when the resource that was asked could not be found.

PUBLIC_ROOMS

GET /_matrix/client/v3/publicRooms

REFRESH_TOKEN

POST /_matrix/client/v3/refresh without new refresh token.

REFRESH_TOKEN_WITH_REFRESH_TOKEN

POST /_matrix/client/v3/refresh with a new refresh token.

REGISTRATION_RESPONSE_ERR

Failed call to POST /_matrix/client/v3/register

UNKNOWN_TOKEN_SOFT_LOGOUT

Failed called to any endpoint with an expired access token.

VERSIONS

GET /_matrix/client/versions

WELL_KNOWN

GET /.well-known/matrix/client

WHOAMI

GET /_matrix/client/v3/account/whoami