Crate matrix_sdk_test

Modules

appservice

test_json

Test data for the matrix-sdk crates.

Structs

EventBuilder

The EventBuilder struct can be used to easily generate valid sync responses for testing. These can be then fed into either Client or Room.

Enums

EventsJson

Embedded event files

SyncResponseFile

Embedded sync response files

Functions

response_from_file

sync_response

Get specific API responses for testing

Attribute Macros

async_test

Attribute to use wasm_bindgen_test for wasm32 targets and tokio::test for everything else