Module matrix_sdk_test::test_json

Test data for the matrix-sdk crates.

Exporting each const allows all the test data to have a single source of truth. When running cargo publish no external folders are allowed so all the test data needs to be contained within this crate.

Re-exports

pub use api_responses::DEVICES;

pub use api_responses::GET_ALIAS;

pub use api_responses::KEYS_QUERY;

pub use api_responses::KEYS_UPLOAD;

pub use api_responses::LOGIN;

pub use api_responses::LOGIN_RESPONSE_ERR;

pub use api_responses::LOGIN_TYPES;

pub use api_responses::LOGIN_WITH_DISCOVERY;

pub use api_responses::PUBLIC_ROOMS;

pub use api_responses::REGISTRATION_RESPONSE_ERR;

pub use api_responses::VERSIONS;

pub use api_responses::WELL_KNOWN;

pub use api_responses::WHOAMI;

pub use members::MEMBERS;

pub use messages::ROOM_MESSAGES;

pub use messages::ROOM_MESSAGES_BATCH_1;

pub use messages::ROOM_MESSAGES_BATCH_2;

pub use sync::DEFAULT_SYNC_ROOM_ID;

pub use sync::DEFAULT_SYNC_SUMMARY;

pub use sync::INVITE_SYNC;

pub use sync::LEAVE_SYNC;

pub use sync::LEAVE_SYNC_EVENT;

pub use sync::MORE_SYNC;

pub use sync::MORE_SYNC_2;

pub use sync::SYNC;

pub use sync::VOIP_SYNC;

pub use sync_events::ALIAS;

pub use sync_events::ALIASES;

pub use sync_events::ENCRYPTION;

pub use sync_events::MEMBER;

pub use sync_events::MEMBER_INVITE;

pub use sync_events::MEMBER_NAME_CHANGE;

pub use sync_events::MEMBER_STRIPPED;

pub use sync_events::MESSAGE_EDIT;

pub use sync_events::MESSAGE_TEXT;

pub use sync_events::NAME;

pub use sync_events::NAME_STRIPPED;

pub use sync_events::POWER_LEVELS;

pub use sync_events::PRESENCE;

pub use sync_events::PUSH_RULES;

pub use sync_events::REACTION;

pub use sync_events::READ_RECEIPT;

pub use sync_events::READ_RECEIPT_OTHER;

pub use sync_events::REDACTED;

pub use sync_events::REDACTED_INVALID;

pub use sync_events::REDACTED_STATE;

pub use sync_events::REDACTION;

pub use sync_events::TAG;

pub use sync_events::TOPIC;

pub use sync_events::TYPING;

Modules

api_responses

Responses to client API calls.

members

Example responses to GET /_matrix/client/v3/rooms/{roomId}/members

messages

Example responses from calls to GET /_matrix/client/v3/rooms/{roomId}/messages.

sync

Complete sync responses.

sync_events

Discrete events found in a sync response.

Statics

EMPTY

An empty response.

EVENT_ID

A response with only an event ID.

ROOM_ID

A response with only a room ID.