Module matrix_sdk_test::test_json

Test data for the matrix-sdk crates.

Exporting each const allows all the test data to have a single source of truth. When running cargo publish no external folders are allowed so all the test data needs to be contained within this crate.

Re-exports

pub use events::ALIAS;

pub use events::ALIAS;

pub use events::ALIASES;

pub use events::ALIASES;

pub use events::EVENT_ID;

pub use events::EVENT_ID;

pub use events::KEYS_QUERY;

pub use events::KEYS_QUERY;

pub use events::KEYS_UPLOAD;

pub use events::KEYS_UPLOAD;

pub use events::LOGIN;

pub use events::LOGIN;

pub use events::LOGIN_RESPONSE_ERR;

pub use events::LOGIN_RESPONSE_ERR;

pub use events::LOGOUT;

pub use events::LOGOUT;

pub use events::MEMBER;

pub use events::MEMBER;

pub use events::MEMBER_NAME_CHANGE;

pub use events::MEMBER_NAME_CHANGE;

pub use events::MESSAGE_EDIT;

pub use events::MESSAGE_EDIT;

pub use events::MESSAGE_TEXT;

pub use events::MESSAGE_TEXT;

pub use events::NAME;

pub use events::NAME;

pub use events::POWER_LEVELS;

pub use events::POWER_LEVELS;

pub use events::PRESENCE;

pub use events::PRESENCE;

pub use events::PUBLIC_ROOMS;

pub use events::PUBLIC_ROOMS;

pub use events::REACTION;

pub use events::REACTION;

pub use events::REDACTED;

pub use events::REDACTED;

pub use events::REDACTED_INVALID;

pub use events::REDACTED_INVALID;

pub use events::REDACTED_STATE;

pub use events::REDACTED_STATE;

pub use events::REDACTION;

pub use events::REDACTION;

pub use events::REGISTRATION_RESPONSE_ERR;

pub use events::REGISTRATION_RESPONSE_ERR;

pub use events::ROOM_ID;

pub use events::ROOM_ID;

pub use events::ROOM_MESSAGES;

pub use events::ROOM_MESSAGES;

pub use events::TYPING;

pub use events::TYPING;

pub use sync::DEFAULT_SYNC_SUMMARY;

pub use sync::DEFAULT_SYNC_SUMMARY;

pub use sync::INVITE_SYNC;

pub use sync::INVITE_SYNC;

pub use sync::LEAVE_SYNC;

pub use sync::LEAVE_SYNC;

pub use sync::LEAVE_SYNC_EVENT;

pub use sync::LEAVE_SYNC_EVENT;

pub use sync::MORE_SYNC;

pub use sync::MORE_SYNC;

pub use sync::SYNC;

pub use sync::SYNC;

pub use sync::VOIP_SYNC;

pub use sync::VOIP_SYNC;

Modules

events

sync

Structs

DEVICES